+++ stock-api/stock/api.py
@@ -1,7 +1,6 @@
+from data_fetcher import DataFetcher
 from flask import Flask
 from flask_caching import Cache
-
-from services.data_fetcher import DataFetcher
 
 
 app = Flask(__name__)
@@ -12,7 +11,7 @@ data_fetcher = DataFetcher()
 @app.route("/v1/stock/all/stocks")
 @cache.cached(timeout=60)
 def get_all_stocks():
-	return data_fetcher.get_all_stocks()
+    return data_fetcher.get_all_stocks()
 
 
 @app.route("/v1/stock/analysis/<stock>")
