+++ stock-api/stock/dupont_service.py
@@ -1,4 +1,4 @@
-from services.metric_names import MetricNames
+from metric_names import MetricNames
 from singleton import Singleton
 from stores import FinancialStatementEntryStore
 from time_series import TimeSeries
@@ -10,12 +10,12 @@ class DupontService():
     store = FinancialStatementEntryStore()
     metric_names = MetricNames()
 
-    def get_roe(self, stock):
+    def roe(self, stock):
         net_profit_list = self.store.get(self.metric_names.get('NetProfit'))
         print net_profit_list
         # Yearly and Quarterly
 
         return "FinancialStatementEntryService - ROE"
 
-    def get_equity_multiplier(self, stock):
+    def equity_multiplier(self, stock):
         return "FinancialStatementEntryService - Equity Multiplier"
