+++ stock-api/stock/singleton.py
