+++ stock-api/stock/tests/unit/test_dupont_service.py
@@ -1,0 +1,10 @@
+from dupont_service import DupontService
+
+import mock
+import unittest
+
+
+#@mock.patch('dupont_service.DupontService')
+class DupontServiceTest(unittest.TestCase):
+    def test_roe(self):
+        pass
